Added our webpages for the Project.
These Webpages according to the CA Description are going to be linked to the website.html. This will be my last commit so I can finish working on the wireframe, I am not sure if we need to share the same CSS and folders or use our own individual images and css folders so I might email Hamilton and tell about this.